Fix size and update test.
